feat: display currently playing song status in CLI

+++ player.js
@@ -9,6 +9,7 @@ const songs = fs
 let selected = 1;
 let player = null;
 let isPaused = false;
+let currSong = null;
 
 showSongs();
 
@@ -31,6 +32,7 @@ process.stdin.on("data", (input) => {
     if (player) {
       player.kill();
     }
+    clearCurrent();
     console.log("Thanks for using Mplayer!");
     process.stdin.setRawMode(false);
     process.exit(0);
@@ -40,7 +42,8 @@ process.stdin.on("data", (input) => {
   if (input[2] === "A") {
     if (selected === 1) return;
     selected--;
-    process.stdout.write(`\x1b[${songs.length}A`);
+    const linesToMove = currSong ? songs.length + 1 : songs.length;
+    process.stdout.write(`\x1b[${linesToMove}A`);
     showSongs();
   }
 
@@ -48,19 +51,24 @@ process.stdin.on("data", (input) => {
   if (input[2] === "B") {
     if (selected === songs.length) return;
     selected++;
-    process.stdout.write(`\x1b[${songs.length}A`);
+    const linesToMove = currSong ? songs.length + 1 : songs.length;
+    process.stdout.write(`\x1b[${linesToMove}A`);
     showSongs();
   }
 });
 
 function showSongs() {
-  process.stdout.write("\x1b[2K");
   for (let i = 0; i < songs.length; i++) {
+    process.stdout.write("\x1b[2K\r");
     if (selected === i + 1) {
       console.log(`→ ${i + 1}: ${songs[i]}`);
     } else {
       console.log(`  ${i + 1}: ${songs[i]}`);
     }
+  }
+  if (currSong) {
+    process.stdout.write("\x1b[2K\r");
+    console.log(`Currently Playing: ${currSong}`);
   }
 }
 
@@ -68,17 +76,37 @@ function playSong() {
   if (player) {
     player.kill();
   }
+  clearCurrent();
   const songPath = path.join(__dirname, "Songs", songs[selected - 1]);
+  currSong = songs[selected - 1];
   player = spawn("afplay", [songPath]);
   isPaused = false;
+  consoleCurrent();
 }
 
 function togglePause() {
+  if (!player) {
+    return;
+  }
   if (isPaused) {
     player.kill("SIGCONT");
     isPaused = false;
   } else {
     player.kill("SIGSTOP");
     isPaused = true;
+  }
+}
+
+function consoleCurrent() {
+  if (player && currSong) {
+    process.stdout.write("\x1b[2K\r");
+    console.log(`Currently Playing: ${currSong}`);
+  }
+}
+
+function clearCurrent() {
+  if (currSong) {
+    process.stdout.write("\x1b[1A\x1b[2K\r");
+    currSong = null;
   }
 }
